worker_requirements.txt
azure-servicebus==7.6.0
azure-cosmos==4.2.0